requirements.txt
google-generativeai
gradio
pillow
python-dotenv
rembg
onnxruntime
